requirements.txt
colorama==0.4.6
hydra-core==1.3.2
icecream==2.1.3
loguru==0.7.2
numpy==2.3.1
omegaconf==2.3.0
pandas==2.3.1
Pillow==11.3.0
regex==2024.11.6
scikit_learn==1.7.0
torch==2.7.0
torchmetrics==1.4.2
tqdm==4.67.1
transformers==4.52.4
wandb==0.21.0
